I downloaded the auto-generated server from the swagger-editor
This code was generated by swagger codegen. Codegen generates code for various
languages. In prior projects I've worked on, the openapi-generate-cli was used.
The layout is similar:

api_*.go
  These are the handlers for the http requrests. The internals of these methods
  must be modified.

logger.go
  A middle layer to log http requests after they've been handled

model_*.go
  The data structures defined by the schemas of the open api spec

routers.go
  Leverages gorilla mux to assign the handlers to the http endpoint according to
  the http method. This is all done in the NewRouter function. The logger
  middleware is set here for all of the endpoints. The function doesn't use
  gorilla's subrouter abstraction - each route struct is defined with its
  complete path hardcoded.  Having the full api path makes it easy to determine
  its mapping to the handler.  Gorilla's middleware abstraction also isn't used.
  The logger middle layer is set by manually wrapping each of the handlers
  before they are registered. This part of the code is more imperative (it
  requires you to understand the procedure defined in NewRouter, this could
  easily be refactored to be more declaritive by having a middleware member in
  the route structure.

  Registering an endpoint is 10 lines of procedural code, the other 190 lines
  are defining the values of the routes in a structured, declarative way.

main.go
  Gets the routes (from the NewRouter function) and runs the
  http server. Short and sweet.

Procedures are succinctly defined. The bulk of the code declares the data that
the procedure will work on. Because the data and code strcture follow a clear
pattern, I don't have to read every line of every function to understand how the
program will behave!

Granted this doesn't do anything yet. But I think the code is really nicely
organized. I know which parts of the codebase that I need to expand on.

I also need to be able to build it. The package structure of the autogenerated
code doesn't fit go's module conventions.

If the OpenAPI spec for the project is modified, this code will have to be
regenerated? That's another area that requires some thought.