It finally builds!
The codebase has been initialized as a module, and the external packages
are now tracked in go.sum.

A Makefile was created with targets to build and cleanup.

gitignore was added to ignore the binary artifacts that will be made.

routers.go was changes to routes.go, since that is what is ultimatley
what is most import in that file.

Now I can actually do work on it...