Begin implementation of pet service endpoints.
Pet is the easiest endpoint to start iwth, as it is the least dependent
on the other entities.

The handler is provided an http Request and an http Response object. The
pet object must be extracted from the body of the request. The http Body
field is a reader/closer, so it has the method:
      Read(p []byte) (n int, err error)
This will give us a byte array with the pet in json notation. This data
then needs to be unmarshalled to get the golang pet object.

Once the pet object has been extracted, it is time to actually do
something with the data. As a simple first step, the pet object is
placed into an in-memory array. Its id is set to be its index within
this array. Simplest implementation of Add Pet.

While small, the service logic isn't in its correct place: it was put in
a handler file. This will be refactored later- you shouldn't see service
logic mixed in with the handler logic. It should be encapsulated.

After the service logic runs, the handler needs to write the http
response. On successful runs, the pet object is marshalled into json and
written into the response body. If an error occurred, the response
header is written to reflect this error.