Get pet by id.
The url variable is retreived by calling mx.Vars(request), which will
return a mapping of the variables to their string values.

If these values are attained in the handler, that means the handler will
also depend on mux. The router package also depends on this. There are 5
packages in this codebase (main, router, handlers,models, middleware).
The packages router and handlers share external package dependencies
like mux and internal packages such as net/http. Each package in the
codebase is responsible for a different task in the code, the packages
it uses are imported to accomplish that task. The dependencies between
the codebase packages are important, but their use of external packages
are not. Although you can get a sense for what the package will do based
on the packages it imports!

AddPet initially wasn't reading the request body because the byte array
had no capacity. Using ioutil.readAll ensures the full body is read to.
io.ReadAll shows the logic needed to fill and expand a byte array.
Actually, according to the documentation of ioutil, the package io
should be preferred, since ioutil is just a caller for that.

This issue was solved by searching google and comming accros the
"golang-nuts" google group. I subscribed to the group. There were two
takeaways here:
- Forums are a great way to resolve problems if someone has had the same
  issue as you. Otherwise, you have the opportunity to expand the
  group's corpus by asking a question about your specific issue.
- Reading documentation helps you understand the methods in the package
  you are calling and can provide you with helpful tidbits.

On getting the pet, the id needs to start on 1, oherwise it is omitted
from the json body... unless the 'omitempty' tag is removed from the
generated struct.

That part of the code was generated, so there is hesitation to change is
because it may be difficult to regenerate it and have to do all of the
manual changes again. This highlights the tradeoff of generated code.

Generated code is inflexible. In this case, the policy will be to
generate once, then manually modify to fit the needed behavior. Future
regenerations will need to be diffed against this tracked version of the
file to see what chunks should be saved. Future changes should be small,
so there shouldn't be too much work to move. That is my assumption.

Testing the newly added functionality was done manually. For quick
testing curl calls was used. These calls were build through reading the
man page for curl. Since the implementation is simple, and I have full
familiarity, the small manual test is sufficient for me to know that the
server behaves as expected.

I will later want to be able to quickly verify that changes did not
modify behavior. This is immediatley needed to refactor the handler by
extracting out the service logic. In preparation for this refactoring I
added comments tagging what I view to be logic belonging to the
service layer.

Here is the manual test output:
VVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVV
$ cat pet.json
{
  "name": "Brady"
}
$ curl -w "\nHTTP CODE: %{http_code}\n" http://localhost:8080/api/v3/pet/0

HTTP CODE: 404
$ curl http://localhost:8080/api/v3/pet --data @pet.json ; echo
{"id":0,"name":"Brady","photoUrls":null}
$ curl http://localhost:8080/api/v3/pet --data @pet.json ; echo
{"id":1,"name":"Brady","photoUrls":null}
$ curl -w "\nHTTP CODE: %{http_code}\n" http://localhost:8080/api/v3/pet/-1

HTTP CODE: 404
$ curl -w "\nHTTP CODE: %{http_code}\n" http://localhost:8080/api/v3/pet/0
{"id":0,"name":"Brady","photoUrls":null}
HTTP CODE: 200
$ curl -w "\nHTTP CODE: %{http_code}\n" http://localhost:8080/api/v3/pet/1
{"id":1,"name":"Brady","photoUrls":null}
HTTP CODE: 200
$ curl -w "\nHTTP CODE: %{http_code}\n" http://localhost:8080/api/v3/pet/2

HTTP CODE: 404
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^