Added some more tests. Took a while.
These tests are simple, but they capture the simple behavior that the
backend server exhibits so far:
- 404 is returned when the call tries to get a pet that doesn't exist
- When a pet is created, 200 is returned and the body has the pet data
- When the pet does exist, 200 is return and the body has the pet data

The test script expects that a fresh instance of the backend is running.
Trying to run the test a second time without a clean instance of the
backend will fail tests.

Later I'll look into making the test start up the webserver. I could do
it all in containers. That would be pretty slick.

But I need to add some new functionality to this backend. This is
starting to become one of those projects that I don't actually work on.
Probably because the spec has no real bearing on anything I'm actually
doing.

This is an exercise of commitment.