Delete pet endpoint added!
Finally some true backend development work.

My dedication to study has not been consistent. But it doesn't need to
be right off-the-bat. I learn little by little. If I don't rush, I'll
get the full value of what I am doing. Slowly the bad habits will fade,
and the principles that work through quotodial practice will shine
through! That's the idea anyways...

Here's a bad habit: drinking alcohol. Here's another bad habit:
overthinking. "Analysis paralysis". I started today's session with "The
petstore api has a bunch of fields that I don't think are all that
useful". Then I realized: I haven't done anything productive on this
project in weeeeeeks. The idea was to get good at building backends.
Instead I sit and criticize.

My thinking going in:
Focus. Start small. Build on it with future backends. In this exercise,
just take the api and make a functioning backend server. Ignore data
persistance. Ignore tests. Ignore refactoring. Ignore security. Let's
get it done, then improve later. I can't start lifting with 200 lbs.

Well, I ended up writing tests for delete. I think it is good practice
to start by writing the integration tests first. This starts the
thinking by defining the actual behavior that we want to see, without
thinking about how the implementation will do it. It's a good starting
mindset, because this is how the user will interact with it. Simple
stuff:
- On successful delete: 204
- If the pet doesn't exist: 404

Onto the implementation. That's when I realized the data structure I'm
using does not handle deletes well. The id is based on the list index,
so deleting a pet shifts all of the ids. I changed this to a map and a
generator function to create the ids.

This required some changes to the create and get handlers. The previous
test I made came in handy, since I was able to use that to quickly
verify that the server still exhibits the desired behavior.

The behavior for error handling done by the handler depends on the error
that was received from the service code. If the per wasn't found, 404
should be returned.  Other errors should be logged, and 500 (internal
server error) should be returned. The errors package provides a way to
create and compare errors in this way.

The tests still require a clean slate before they are run. That's ok. My
focus is the endpoint functionality. 3/8 pet endpoints done. There are
19 endpoints total.