pet findByStatus endpoint.
Piece of cake with the write tests! The example OpenAPI spec defined an
enum for the status of the pets: {available, pending, sold}. I was
suprised to find that the code generator didn't automatically generate
an enum set for this field.

Note: At this time, there are no checks to make sure that the status
passed by the user is one of the expected values.

How do I provide the status as a query?
```
                     r.URL.Query().Get("<name>")
```
This wasn't very intuitive to me. This was the part that took the most
debugging.