Validation logic for pets being created using higher order procedures!
Problem: The user can creatre a pet with an invalid status.

Concepts:
  models.Pet is a data abstraction representing a real-world concept: a
  pet, or at least the properties of a pet that is important for the
  tasks this program is designed to handle.

  petstore is the program that has defined what models.Pet is made up
  of. We want petstore to decide if a pet is valid or not. We want
  petstore to DECLARE whether a pet is valid or not.

Validating a potentially valid pet yields either a valid pet or an
invalid pet. If we got a valid pet, we should continue with the service
call as normal. If we got an invalid pet, we would want to communicate
the reason why to the user.

This means that the handling of the pet is different based on if it is
valid or not. This can be neatly expressed through a higher-order
procedure that expects a procedure for if the pet as valid, and one for
if the pet is invalid. The invalid pet will have a list of violations
that was populated when the pet was validated.

Once again, creating some initial tests that defined the behavior of the
server made quick work of this refactor.

VVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVV

[rcampos@rh-laptop test]$ curl \
    --data '{"name": "Ribbons", "status": "rock"}' \
    -w "\nHTTP CODE: %{http_code}\n" http://localhost:8080/api/v3/pet

["\"rock\" is not a valid status"]
HTTP CODE: 400

^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^