Pascal's triangle.
What does this have to do with the pet store? Nothing.
I've deviated from the backend development practice for a fun computing
exercise. I think the time was well spent. It took me longer than I
would have liked to formulate a recursive procedure that computes the
elements of Pascal's triangle, and then more time fiddling with Python's
string formatting mechanisms to get the triangle to print exactly how I
wanted.

It was worth it. I learned some new things, and I accomplished what I
set out to do. Visualize; Plan; Execute.

Here's the trick to making the recursive procedure:
Put Pascal's triangle on a matrix:
  | 0 1 2 3 4
--+-----------
 0| 1 - - - -
 1| 1 1 - - -
 2| 1 2 1 - -
 3| 1 3 3 1 -
 4| 1 4 6 4 1

 Now the set of base cases becomes clear:
 * If row ==   0 -> 1
 * If row == col -> 1
 * If col ==   0 -> 1

When the base case isn't met, a set of recursive calls are made.
It is this recursive expression that describes the pattern of the
triangle: The summation of the two elements directly above the one we're
interested in.

To pretty-print, I used Python's newish string formatting mechanism. I
learned some new concepts about python. When some God-tier developer
wants to make an enhancement to Python, they create a PEP document. PEP
stands for Python Enhancement Proposal. The original PEP for the
formatting syntax was PEP 3101, and is dated April-16-2006.The
documentation lists which PEP introduced these changes.

I tried to read as much documentation as I could, but once I found what
I thought I needed I got too excited and jumped back into my script. It
would have been better to get a complete picture of the format string
before jumping back in the code? I don't know. On one hand, I'm not an
expert on string formatting. On the other, I know where to look, and can
build that skillset as I go. Actually, this was the right way. Reading
the full documentation wouldn't have made me an expert in string
formatting. The experience of using it successfully, however, has
brought me one step close towards proficiency in it.

That was the whole point of this entire exercise: It isn't enough to
just read "Structure and Interpretation of Computer Programs", you have
to do the exercises to WORK towards mastery.

One last note: Donald Knuth's The Art of Computer Programming appears as
reference for both SICP and "Data Structures, An Advanced Approach Using
C". This must be a foundational text if so many of the highly-used
textbooks of today use it as reference. It was published in 1973.

My next move should be to get findPetsWithTags operational, but what I'm
actually going to do is figure out how to animate the creation of this
triangle with Manim... The next commit will reveal the path I chose...
or no path at all. No, there will be another commit. I'm committed to
the commit.