Implemented findPetByTags; Improved findPetByStatus
Looking more closely at the API spec, findByStatus can have multiple
values too.

findPetsByStatus is a Logical Disjunction. The endpoint should return
any pet that has a status matching one in the list of the query.

findPetsByTags is a Logical Conjunction. The endpoint should return any
pet that has all of the listed tags in the query.

There is a recursive procedure for finding pets with a conjunction of
tags:
  pets_with_tags = pets_with_tag(tag1,
                     pets_with_tag(tag2,
                       ...
                          pets_with_tag(tagN, allPets)...))

Using this procedure is luxuriously inefficient compared to the
iterative way I initially came up with... or is it?

Iterative:
2023/01/21 15:30:30 GET /api/v3/pet/findByTags?tags=afc-certified FindPetsByTags 27.493µs
2023/01/21 15:30:30 GET /api/v3/pet/findByTags?tags=foobar FindPetsByTags 16.667µs
2023/01/21 15:30:30 GET /api/v3/pet/findByTags?tags=good_boy,backend_expert FindPetsByTags 21.247µs
2023/01/21 15:30:30 GET /api/v3/pet/findByTags?tags=good_boy,foobar FindPetsByTags 13.429µs

Recursive:
2023/01/21 15:25:49 GET /api/v3/pet/findByTags?tags=afc-certified FindPetsByTags 15.051µs
2023/01/21 15:25:49 GET /api/v3/pet/findByTags?tags=foobar FindPetsByTags 19.373µs
2023/01/21 15:25:49 GET /api/v3/pet/findByTags?tags=good_boy,backend_expert FindPetsByTags 25.697µs
2023/01/21 15:25:49 GET /api/v3/pet/findByTags?tags=good_boy,foobar FindPetsByTags 4.083µs

It is just one sample, but the numbers don't look all that different.
Maybe with a large number of tags or pets this would matter... But how
many pets will this store hold? I can't imagine more than a couple
thousand in its lifetime, if that.

I'll keep the recursive appraoch, since it is more expressive.

Once more endpoint down. Three more to go for pet.