updating pet by form with form data
What is meant by form data? Through URL parameters, also known as a
Query String. From Wikipedia:

    A query string is a part of a uniform resource locator that assigns
    values to specified parameters.

Here is an example:

http://localhost:8080/api/v3/pet/0?status="Available"\&name="Rocky"

What are the tradeoffs between using URL parameters instead of a JSON
body to update a resource? Here is it clear what needs to be updated and
what shouldn't be. If the user only wants to update the status, the
parameter name is omitted from the URL. This approach only  works for
updating simple fields. For resources with objects within them, this
approach would require a way to access the inner object as a
subresource.  A json body would make experessing this type of change
easy without a need to point to the specific subresource.

For testing, this is the new behavior that needed to be handled:
- updating name
- updating status
- updating both
---> All return 200 and a json body with the updated pet
- invalid update
---> All return 400 and a json body listing the request violations.