Started implementing the uploadPhoto endpoint.
I've been focused on math, writing and getting through SICP, so I
haven't done much in terms of backend development. I haven't been in the
mood to deal with software engineering problems. Am I a software
engineer? Perhaps in name only. My true love is the theoretic. Creating
a procedure that will reverse a list won't put food on the table, but
figuring it out feels like cake for my brain.

(define (reverse seq)
  (accumulate (lambda (x y)
                (if (null? y) (list x)
                    (append y (list x)))) null seq))

To make some forward progress I did a quick implementation that just
writes the uploaded file to a hard-coded location: /tmp/photo.png. On
manually testing, it looked like the photo wasn't being sent properly.
This is the curl command that I was using at the time:
```
curl localhost:8080/api/v3/pet/0/uploadImage -H "Content-Type:application/octet-stream" --data @dog.png
```

How can I know that the file saved matches the file that was uploaded? I
thought about this for a couple of days until it hit me: checksums! It
is common practice to create sha256 hashes of files. If the files are
identical, their hashes will be identical too! I chose sha256 because it
is highly unlikely that there will be any sort of collision by reaching
this algorithm. A collision here means two different files generating
the same hash. I can't prove that a collision will not occur, but I
assume that this is true because sha256 is the algorithm Bitcoin uses
for its proof-of-work, and in the billions of hashes created, no one has
reported a duplicate so far.

The backend response to a successful uploadImage call with the sha256
checksum of the file it has written. The client can check this sum
against its own generated checksum of the file. If the checksums are
identical, the client can be confident that the file was saved correctly
on the backend!

I made a nice little python script that automates this test an it worked
nicely.

The curl command above still doesn't work. That is a mystery for another
day. I also need to enhance the handler so that photos aren't just saved
to /tmp/photo.png. How will I determine the filename? Could I put it in
the http header? Maybe I should consult the spec. The checksum isn't
part of the spec, but I'm keeping it.

The picture used to test was generated using Dall-E 2: An AI image
generator. What a time to be alive! My query was "cute dog handstand". I
very much hope to get back to AI research someday... But I wonder if
I would be in it for the hype, or if it is a genuine interest. I think a
little of both. I think getting too caught on the hype, and losing focus
on what was important to learn is what ultimatley caused me to give up.
When I am ready, I'll put ego aside and study what I need to, so that I
can become a researcher in the field I feel pure passion for.